Move services back to /app folder + UI component fixes
