+++ quick-start/src/main/ui/app/entities/entities.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, EventEmitter } from '@angular/core';
 import { Http, Response } from '@angular/http';
-import { ProjectService } from '../shared/services/projects';
+import { ProjectService } from '../projects';
 import { Subject } from 'rxjs/Subject';
-import { SettingsService } from '../shared/services/settings';
+import { SettingsService } from '../settings';
 
 import { Entity } from './entity.model';
 import { Flow } from './flow.model';
