+++ quick-start/src/main/ui/app/entity-modeler/entity-modeler.component.ts
@@ -10,7 +10,7 @@ import { Connection, Entity, PropertyType } from '../entities';
 
 import { EntitiesService } from '../entities/entities.service';
 
-import { InstallService } from '../shared/services/installer';
+import { InstallService } from '../installer';
 
 import { MdlDialogService, MdlSnackbarService } from '@angular-mdl/core';
