+++ quick-start/src/main/ui/app/installer/index.ts
