+++ quick-start/src/main/ui/app/installer/install.service.ts
