+++ quick-start/src/main/ui/app/jobs/job-listener.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, EventEmitter } from '@angular/core';
 import { Message } from 'stompjs/lib/stomp.min';
 import { STOMPService } from '../stomp';
-import { FlowStatus } from '../../../entities/flow-status.model';
+import { FlowStatus } from '../entities/flow-status.model';
 
 @Injectable()
 export class JobListenerService {
