+++ quick-start/src/main/ui/app/settings/index.ts
