+++ quick-start/src/main/ui/app/settings/settings.service.ts
