+++ quick-start/src/main/ui/app/shared/components/jobs/jobs-ui.component.ts
@@ -1,7 +1,7 @@
 import {ChangeDetectionStrategy, Component, EventEmitter, Input, Output} from "@angular/core";
 import { SearchResponse } from '../../../search';
 import { Job } from './job.model';
-import { JobListenerService } from "../../services/jobs/";
+import { JobListenerService } from "../../../jobs/";
 import { differenceInSeconds } from 'date-fns';
 
 @Component({
